differentiated int vs double primitive type

+++ Java/declaration.java
@@ -3,7 +3,7 @@ public class declaration {
     public static void main(String[] args) {
         int a = 10;
         int b = 5;
-        int c = a + b;
+        double c = a + b;
         System.out.println( a + " + " + b + " = " + c);   
     }
